fix: extract shared LogsDrawer, add Escape key, responsive stats grid
- Extract duplicated LogsDrawer into shared component with Escape key dismiss
- Update both Deployments.jsx and Platform.jsx to use the shared component
- Make Deployments stats grid responsive (grid-cols-2 sm:grid-cols-4)
