+++ frontend/src/pages/Deployments.jsx
@@ -19,7 +19,6 @@ import {
   CheckCircle2,
   Circle,
   ExternalLink,
-  X,
 } from 'lucide-react'
 
 import {
@@ -27,11 +26,11 @@ import {
   startDeployment,
   stopDeployment,
   restartDeployment,
-  getDeploymentLogs,
 } from '../services/api'
 import { useToast } from '../components/Toast'
 import PageHeader from '../components/shared/PageHeader'
 import EmptyState from '../components/shared/EmptyState'
+import LogsDrawer from '../components/shared/LogsDrawer'
 
 const POLL_MS = 5000
 
@@ -92,49 +91,6 @@ const IconBtn = ({ icon: Icon, onClick, title, tone = 'gray', disabled = false }
     >
       <Icon className="w-4 h-4" />
     </button>
-  )
-}
-
-const LogsDrawer = ({ containerName, onClose }) => {
-  const { data, isLoading, refetch, isFetching } = useQuery({
-    queryKey: ['deployment-logs', containerName],
-    queryFn: () => getDeploymentLogs(containerName, 300),
-    enabled: !!containerName,
-  })
-
-  if (!containerName) return null
-
-  return (
-    <div className="fixed inset-0 z-40 flex">
-      <div className="flex-1 bg-black/30" onClick={onClose} />
-      <div className="w-[720px] max-w-[90vw] bg-white h-full flex flex-col border-l border-gray-200 shadow-xl">
-        <div className="flex items-center justify-between px-4 py-3 border-b border-gray-200">
-          <div className="flex items-center gap-2">
-            <FileText className="w-4 h-4 text-gray-600" />
-            <span className="font-medium text-sm">{containerName}</span>
-            <span className="text-xs text-gray-500">last 300 lines</span>
-          </div>
-          <div className="flex items-center gap-2">
-            <button
-              onClick={() => refetch()}
-              className="p-1.5 text-gray-500 hover:text-gray-700 hover:bg-gray-100 rounded"
-              title="Refresh"
-            >
-              <RefreshCw className={`w-4 h-4 ${isFetching ? 'animate-spin' : ''}`} />
-            </button>
-            <button
-              onClick={onClose}
-              className="p-1.5 text-gray-500 hover:text-gray-700 hover:bg-gray-100 rounded"
-            >
-              <X className="w-4 h-4" />
-            </button>
-          </div>
-        </div>
-        <pre className="flex-1 overflow-auto p-4 text-xs font-mono bg-gray-900 text-gray-100 whitespace-pre-wrap">
-{isLoading ? 'Loading…' : (data?.logs || '(no logs)')}
-        </pre>
-      </div>
-    </div>
   )
 }
 
@@ -259,7 +215,7 @@ const Deployments = () => {
       </PageHeader>
 
       {/* Stats row */}
-      <div className="grid grid-cols-4 gap-3">
+      <div className="grid grid-cols-2 sm:grid-cols-4 gap-3">
         <StatCard label="Total" value={stats.total} />
         <StatCard label="Running" value={stats.running} tone="green" />
         <StatCard label="Stopped" value={stats.stopped} tone="gray" />
